+++ plugins/main/public/components/common/dashboards/dashboard.tsx
@@ -4,17 +4,14 @@ import {
   withErrorBoundary,
   withInjectProps,
 } from '../hocs';
-import { getPlugins } from '../../../kibana-services';
 import { LoadingSearchbarProgress } from '../loading-searchbar-progress/loading-searchbar-progress';
 import { useReportingCommunicateSearchContext } from '../hooks/use-reporting-communicate-search-context';
 import { WzSearchBar } from '../search-bar';
 import { DiscoverNoResults } from '../no-results/no-results';
 import { SampleDataWarning } from '../../visualize/components';
 import { compose } from 'redux';
-import { ViewMode } from '../../../../../../src/plugins/embeddable/public';
-
-const DashboardByRenderer =
-  getPlugins().dashboard.DashboardContainerByValueRenderer;
+import DashboardRenderer from './dashboard-renderer/dashboard-renderer';
+import classnames from 'classnames';
 
 export const Dashboard = props => {
   // This is not used by the SCA dashboard.
@@ -29,6 +26,10 @@ export const Dashboard = props => {
       to: props.dataSource.searchBarProps.dateRangeTo,
     },
   });
+
+  const shouldHideDashboard = !Boolean(
+    props.dataSourceAction?.data?.hits?.total > 0,
+  );
 
   return (
     <>
@@ -45,59 +46,34 @@ export const Dashboard = props => {
       {props.dataSourceAction?.data?.hits?.total === 0 ? (
         <DiscoverNoResults />
       ) : null}
-      <div
-        className={
-          props.dataSourceAction?.data?.hits?.total > 0 ? '' : 'wz-no-display'
-        }
-      >
+      <div>
         {props.sampleDataWarningCategories && (
-          <SampleDataWarning
-            categoriesSampleData={props.sampleDataWarningCategories}
-          />
+          <div
+            className={classnames({
+              'wz-no-display': shouldHideDashboard,
+            })}
+          >
+            <SampleDataWarning
+              categoriesSampleData={props.sampleDataWarningCategories}
+            />
+          </div>
         )}
 
         <div className='wz-dashboard-responsive'>
           {props.getDashboardPanels.map(
-            ({
-              getDashboardPanels,
-              id,
-              title,
-              description,
-              hidePanelTitles,
-              useMargins = true,
-              className = '',
-            }) => {
-              const idComponent = id;
+            ({ dashboardId, agentDashboardId, className = '' }) => {
               const dashboard = (
-                <DashboardByRenderer
-                  key={idComponent}
-                  className='wz-search-me'
-                  input={{
-                    viewMode: ViewMode.VIEW,
-                    panels: getDashboardPanels(
-                      props.dataSource.dataSource.id,
-                      Boolean(
-                        props.dataSource.dataSource?.getPinnedAgentFilter()
-                          ?.length,
-                      ),
-                    ),
-                    isFullScreenMode: false,
-                    filters: props.dataSource.fetchFilters ?? [],
-                    useMargins,
-                    id: id,
-                    timeRange: {
-                      from: props.dataSource.searchBarProps.dateRangeFrom,
-                      to: props.dataSource.searchBarProps.dateRangeTo,
-                    },
-                    title: title,
-                    description: description,
-                    query: props.dataSource.searchBarProps.query,
-                    refreshConfig: {
-                      pause: false,
-                      value: 15,
-                    },
-                    hidePanelTitles,
-                    lastReloadRequestTime: props.dataSource.fingerprint,
+                <DashboardRenderer
+                  dashboardId={dashboardId}
+                  agentDashboardId={agentDashboardId}
+                  className={classnames(className, {
+                    'wz-no-display': shouldHideDashboard,
+                  })}
+                  hasPinnedAgent={Boolean(
+                    props.dataSource.dataSource?.getPinnedAgentFilter()?.length,
+                  )}
+                  config={{
+                    dataSource: props.dataSource,
                   }}
                 />
               );
@@ -106,7 +82,10 @@ export const Dashboard = props => {
                 /* Add a wrapper div with the className to apply styles that allow to overwrite
                 some styles using CSS selectors */
                 return (
-                  <div className={className} key={idComponent}>
+                  <div
+                    className={className}
+                    key={dashboardId || agentDashboardId}
+                  >
                     {dashboard}
                   </div>
                 );
@@ -139,15 +118,8 @@ export const createDashboard = ({
   DataSourceRepositoryCreator: any;
   sampleDataWarningCategories?: string[];
   getDashboardPanels: Array<{
-    getDashboardPanels: (
-      indexPatternId: string,
-      isPinnedAgent?: boolean,
-    ) => any;
-    id: string;
-    title: string;
-    description: string;
-    hidePanelTitles: boolean;
-    useMargins?: boolean;
+    dashboardId: string;
+    agentDashboardId?: string;
     /**
      * Class name to apply to the dashboard container
      */